CF-611: Permission error at first time bootstrap
Terraform sometimes ignores the dependencies between ressources, for example when it comes to the Workload Identity Federation ressources. The Service Account may misses the permissions to create the resources, and Terraform fails. To bypass this we do run a targeted apply at bootstrap, ensure the service account has all needed permissions.