Use standard storage class instead of nearline for tf state bucket (#38)
standard storage class seems to be a more suitable choice for tf bucket, as the content can change quite often and is not really intended to be archived for long time, that would make nearline preferable. So, change the hardcoded class to standard. Possible future improvement: make it configurable via parameter.