feat: dns logging policy
If VPC creation is enabled the module will automatically create a DNS policy to enable logging.
As you can only attach one DNS policy to a VPC, the creation of the policy can be disabled. This
will allow users to create a custom policy and attach it to the VPC, if needed.